1.6.0.8 - Turns out other rows in the static data dump had changed so checked and reimported everything.

+++ trunk/DatabaseClasses/Items.cs
@@ -305,7 +305,7 @@ namespace EveMarketMonitorApp.DatabaseClasses
             }
             else
             {
-                short groupID = itemData.groupID;
+                int groupID = itemData.groupID;
                 if (groupTable == null)
                 {
                     groupTable = new EveDataSet.invGroupsDataTable();
